pyproject.toml
[project]
name = "simplechatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "nltk>=3.9.2",
    "numpy>=2.4.2",
    "pinecone>=8.0.1",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.2",
]
